+++ src/main/java/tn/dalhia/services/implementations/CommandServiceImpl.java
@@ -1,10 +1,18 @@
 package tn.dalhia.services.implementations;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.transaction.Transactional;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import tn.dalhia.entities.Command;
 import tn.dalhia.entities.Product;
@@ -14,10 +22,11 @@ import tn.dalhia.repositories.CommandRepository;
 import tn.dalhia.repositories.ProductRepository;
 import tn.dalhia.repositories.UserRepository;
 import tn.dalhia.request.CommandRequestModel;
+import tn.dalhia.request.CommandRequestProducts;
 import tn.dalhia.response.ErrorMessages;
 import tn.dalhia.services.CommandService;
 import tn.dalhia.shared.dto.CommandDto;
-import tn.dalhia.shared.dto.Utils;
+import tn.dalhia.shared.tools.UtilsUser;
 
 @Service
 public class CommandServiceImpl implements CommandService{
@@ -32,22 +41,44 @@ public class CommandServiceImpl implements CommandService{
 	ProductRepository productRepo;
 	
 	@Autowired
-	Utils utils;
+    UtilsUser utils;
 	
 	@Override
 	@Transactional
-	public CommandDto createCommand(CommandRequestModel commandDetails, String id) {
+	public CommandDto createCommand(CommandRequestModel commandDetails, String id, Authentication authentification) {
 		User userEntity = userRepo.findByUserId(id);
-		Product ProductEntity = productRepo.findByProductId(commandDetails.getProductId());
+		List<Product> products = new ArrayList<>();
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
-		if (ProductEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if (commandDetails.getProducts() == null) throw new UserServiceException(ErrorMessages.MISSING_REQUIRED_FILED.getErrorMessage());
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
+		
+		for (CommandRequestProducts commandRequestProducts : commandDetails.getProducts()) {
+			Product testProduct = productRepo.findByProductId(commandRequestProducts.getIdProducts());
+			if (testProduct == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+			if (testProduct.getQuantity() == 0 || testProduct.getQuantity()-commandDetails.getQuantity()<0) throw new UserServiceException(ErrorMessages.QUANTITY_OVER.getErrorMessage());
+			testProduct.setQuantity(testProduct.getQuantity() - commandDetails.getQuantity());
+			productRepo.save(testProduct);
+			products.add(testProduct);
+		}
 		
 		Command commandEntity = new Command();
 		BeanUtils.copyProperties(commandDetails,commandEntity);
+		
+		// ?? add command question 
 		commandEntity.setCommandId(utils.generateCommandId(30));
 		commandEntity.setUsers(userEntity);
-		//commandEntity.getProducts().add(ProductEntity);  //nes2el monsieur
-		
+		commandEntity.setProducts(products);
+//		for (Product products : commandDetails.getProducts()) {
+//			commandEntity.getProducts().add(products);
+//		}
+//		if(commandEntity.getProducts() != null) {
+//		commandEntity.getProducts().add(ProductEntity);}
+//		else{
+//			List<Product> lP = new ArrayList<Product>();//nes2el monsieur
+//			lP.add(ProductEntity);
+//			commandEntity.setProducts(lP);
+//		}
+
 		Command storedCommand = commandRepo.save(commandEntity);
 		ModelMapper modelMapper = new ModelMapper();
 		CommandDto returnValue = modelMapper.map(storedCommand, CommandDto.class);
@@ -55,9 +86,11 @@ public class CommandServiceImpl implements CommandService{
 	}
 
 	@Override
-	public CommandDto updateCommand(CommandRequestModel commandDetails, String id) {
+	public CommandDto updateCommand(CommandRequestModel commandDetails, String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		Command commandEntity = commandRepo.findByCommandId(id);
 		if (commandEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		
 		commandEntity.setCard(commandDetails.getCard());
 		commandEntity.setCode(commandDetails.getCode());
 		commandEntity.setEmail(commandDetails.getEmail());
@@ -71,18 +104,42 @@ public class CommandServiceImpl implements CommandService{
 	}
 
 	@Override
-	public CommandDto getCommandById(String id) {
+	public CommandDto getCommandById(String id, Authentication authentification) {
 		ModelMapper modelMapper = new ModelMapper();
 		Command command = commandRepo.findByCommandId(id);
+		if (command == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,command.getUsers())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		CommandDto returnValue = modelMapper.map(command,CommandDto.class);
 		return returnValue;
 	}
 
 	@Override
-	public void deleteCommand(String id) {
+	public void deleteCommand(String id, Authentication authentification) {
 		Command command = commandRepo.findByCommandId(id) ;
 		if (command == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,command.getUsers())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		commandRepo.delete(command);
+	}
+
+	@Override
+	public List<CommandDto> getCommandsPagination(int page, int limit, Authentication authentification) {
+		List<CommandDto> returnValue = new ArrayList<>();
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
+		if(page>0) page = page-1;
+		
+		Pageable pageableRequest = PageRequest.of(page, limit);
+		Page<Command>  commandsPage= commandRepo.findAll(pageableRequest);
+		
+		List<Command> commands = commandsPage.getContent();
+		
+		for (Command command : commands) {
+			
+			ModelMapper modelMapper = new ModelMapper();
+			CommandDto commandDto = modelMapper.map(command,CommandDto.class);
+			 //BeanUtils.copyProperties(command, commandDto);
+			 returnValue.add(commandDto);
+		}
+		return returnValue;
 	}
 
 }
