+++ src/main/java/tn/dalhia/services/implementations/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package tn.dalhia.services.implementations;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import tn.dalhia.entities.Product;
@@ -11,7 +12,7 @@ import tn.dalhia.request.ProductRequestModel;
 import tn.dalhia.response.ErrorMessages;
 import tn.dalhia.services.ProductService;
 import tn.dalhia.shared.dto.ProductDto;
-import tn.dalhia.shared.dto.Utils;
+import tn.dalhia.shared.tools.UtilsUser;
 
 @Service
 public class ProductServiceImpl implements ProductService{
@@ -19,10 +20,11 @@ public class ProductServiceImpl implements ProductService{
 	@Autowired
 	ProductRepository productRepo;
 	@Autowired
-	Utils utils;
+    UtilsUser utils;
 	
 	@Override
-	public ProductDto createProduct(ProductRequestModel product) {
+	public ProductDto createProduct(ProductRequestModel product, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		Product productEntity = new Product();
 		ProductDto returnValue = new ProductDto();
 		
@@ -34,7 +36,9 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public ProductDto updateProduct(ProductRequestModel product, String id) {
+
+	public ProductDto updateProduct(ProductRequestModel product, String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		ProductDto returnValue = new ProductDto();
 		Product productEntity = productRepo.findByProductId(id);
 		if (productEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
@@ -42,6 +46,7 @@ public class ProductServiceImpl implements ProductService{
 		productEntity.setDescription(product.getDescription());
 		productEntity.setPrice(product.getPrice());
 		productEntity.setTitle(product.getTitle());
+		productEntity.setQuantity(product.getQuantity());
 		
 		Product updateProduct = productRepo.save(productEntity);
 		BeanUtils.copyProperties(updateProduct,returnValue);
@@ -49,7 +54,9 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public ProductDto getProductById(String id) {
+
+	public ProductDto getProductById(String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		ProductDto returnValue = new ProductDto();
 		Product product = productRepo.findByProductId(id);
 		if (product == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
@@ -58,7 +65,8 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public void deleteProduct(String id) {
+	public void deleteProduct(String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		Product product = productRepo.findByProductId(id);
 		if (product == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		productRepo.delete(product);
